test: add tests for create helper functions
